Move ingest DTOs to a separate file
