+++ app/api/api_v1/routers/ingest.py
@@ -1,25 +1,8 @@
 from fastapi import APIRouter, Request, status
-from pydantic import BaseModel
+
+from app.model.ingest import IngestDTO
 
 ingest_router = r = APIRouter()
-
-
-class DocumentIngestDTO(BaseModel):
-    metadata: list[str]
-    events: list[str]
-
-
-class FamilyIngestDTO(BaseModel):
-    name: str
-    summary: str
-    metadata: list[str]
-    events: list[str]
-    documents: list[DocumentIngestDTO]
-
-
-class IngestDTO(BaseModel):
-    corpus_id: str
-    families: list[FamilyIngestDTO]
 
 
 @r.post("/ingest", response_model=str, status_code=status.HTTP_201_CREATED)
